fix: delete credit card

+++ services/credit_card.services.js
@@ -9,19 +9,15 @@ const findCardById = (id)=>{
     return cardData
 }
 const deleteCard =async (id,userID)=>{
-    const card = await findCardById(id)
-    if(card.userID.toString() !== userID){
-        throw new AppError("Không thể xóa thẻ",400)
-    }
-    const data = await CreditCard.deleteOne({_id:id})
+    const data = await CreditCard.deleteOne({_id:id,userID:userID})
     if(data.deletedCount != 1){
-        throw new AppError("Không thể xóa thẻ, vuil òng thử lại",400)
+        throw new AppError("Không thể xóa thẻ, vui lòng thử lại",400)
     }
 }
 const findCardAndUpdate =async (id,body)=>{
     const data = await CreditCard.updateOne({_id:id},body)
     if(data.modifiedCount !== 1){
-        throw new AppError("Không thể xóa thẻ, vui lòng thử lại",400)
+        throw new AppError("Không thể sửa thẻ",400)
     }
 }
 const getCards = async(id)=>{
